% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{dummyfun}
\alias{dummyfun}
\title{Just another dummy function}
\usage{
dummyfun(x, y)
}
\arguments{
\item{x}{a value}

\item{y}{a value}
}
\value{
a data frame
}
\description{
Just another dummy function
}
