% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{sum2}
\alias{sum2}
\title{Just a dummy function}
\usage{
sum2(x, y)
}
\arguments{
\item{x}{a value to add to y}

\item{y}{a valut to add to x}
}
\value{
the sum of x and y
}
\description{
Just a dummy function
}
